Implements seat selection workflow
Adds a multi-step seat selection workflow.
This includes section selection and seat search.

Introduces a new `RowSteps` component for managing the workflow steps.
Uses framer-motion for smooth transitions and animations.

The UI guides the user through selecting a section of the theater
(Platea, Galleria, Palchi) and then searching for available seats
within that section.

The top section displays the current step and the selected section,
allowing the user to go back and change their selection.

The footer provides a link to the development company.